[frontend] package benchmarks: fix diff calc
