+++ frontend/components/ColorizedValue.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	export let value : number;
-	let color = value < 0 ? 'green' : value > 0 ? 'red' : 'gray';
+	let color = +value.toFixed(2) < 0 ? 'green' : value > 0 ? 'red' : 'gray';
 	let sign = value > 0 ? '+' : '';
 
 </script>
