+++ frontend/components/package/PackageBenchmark.svelte
@@ -1,17 +1,10 @@
 <script lang="ts">
 	import {Benchmark} from '/declarations/main/main.did.js';
 	import ColorizedValue from '../ColorizedValue.svelte';
+	import {getMetricDiff, getMetricNum} from '/logic/benchmark-utils';
 
 	export let benchmark : Benchmark;
 	export let otherBenchmark : Benchmark | undefined = undefined;
-
-	function getMetricNum(bench : Benchmark, row : string, col : string, metric : string) : number | undefined {
-		let metricData = bench.metrics.find((m) => m[0] === metric);
-		let rowIndex = bench.rows.indexOf(row);
-		let colIndex = bench.cols.indexOf(col);
-
-		return Number(metricData?.[1]?.[rowIndex]?.[colIndex]);
-	}
 
 	function getMetric(bench : Benchmark, row : string, col : string, metric : string) {
 		let value = getMetricNum(bench, row, col, metric);
@@ -21,18 +14,6 @@
 		}
 		else {
 			return '-';
-		}
-	}
-
-	function getMetricDiff(row : string, col : string, metric : string) : number {
-		let curValue = getMetricNum(benchmark, row, col, metric);
-		let prevValue = getMetricNum(otherBenchmark, row, col, metric);
-
-		if (curValue && prevValue) {
-			return (curValue - prevValue) / prevValue * 100;
-		}
-		else {
-			return 0;
 		}
 	}
 </script>
@@ -62,7 +43,7 @@
 						{#each benchmark.cols as col}
 							<td>
 								{#if otherBenchmark}
-									<ColorizedValue value={getMetricDiff(row, col, 'instructions')}></ColorizedValue>
+									<ColorizedValue value={getMetricDiff(benchmark, otherBenchmark, row, col, 'instructions')}></ColorizedValue>
 								{:else}
 									{getMetric(benchmark, row, col, 'instructions')}
 								{/if}
@@ -90,7 +71,7 @@
 						{#each benchmark.cols as col}
 							<td>
 								{#if otherBenchmark}
-									<ColorizedValue value={getMetricDiff(row, col, 'rts_heap_size')}></ColorizedValue>
+									<ColorizedValue value={getMetricDiff(benchmark, otherBenchmark, row, col, 'rts_heap_size')}></ColorizedValue>
 								{:else}
 									{getMetric(benchmark, row, col, 'rts_heap_size')}
 								{/if}
